feat!: Pre-calculate pod addresses and pre-create DVs
forget comment

+++ src/interfaces/IByzNft.sol
@@ -9,8 +9,8 @@ interface IByzNft is IERC721Upgradeable {
      * @notice Gets called when a full staker creates a Strategy Module
      * @param _to The address of the staker who created the Strategy Module
      * @param _nounce to calculate the tokenId. This is to prevent minting the same tokenId twice.
-     * @return The tokenId of the newly minted NFT (calculated from the number of Strategy Modules already own by the staker and the staker's address)
+     * @return The tokenId of the newly minted NFT (calculated from the number of Strategy Modules already deployed)
      */
-    function mint(address _to, uint256 _nounce) external returns (uint256);
+    function mint(address _to, uint64 _nounce) external returns (uint256);
 
 }
